Renamed the project names and unit testst
