+++ UnitTests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MSTest.TestAdapter" version="2.1.0" targetFramework="net472" />
+  <package id="MSTest.TestFramework" version="2.1.0" targetFramework="net472" />
+</packages>
